Achieved smoother motion by using floats

+++ Leg.h
@@ -15,18 +15,18 @@ public:
   void setAngle(uint8_t servoPin, int offset, int angleVal);
 
   // moves the leg instantly to the desired position. Returns false if a servo couldn't reach target position
-  bool moveTo(int xValue, int yValue, int zValue);
+  bool moveTo(float xValue, float yValue, float zValue);
 
   // returns true if the leg is touching the ground
   [[nodiscard]] bool touchesGround();
 
   // checks whether the specified point (with L, z value) is reachable by the leg. Returns true if reachable
-  inline bool isReachable(float L, int z);
+  inline bool isReachable(float L, float z);
 
-  // getter-functions for the current x-y-z-Values where the point of the leg is at
-  [[nodiscard]] int getXVal();
-  [[nodiscard]] int getYVal();
-  [[nodiscard]] int getZVal();
+  // getter functions for the current x-y-z-Values where the point of the leg is at
+  [[nodiscard]] float getXVal();
+  [[nodiscard]] float getYVal();
+  [[nodiscard]] float getZVal();
 
 private:
   // Reference to servo driver module
@@ -45,12 +45,15 @@ private:
   const uint8_t pinPushButton;
 
   //stores the current position (x, y, z)
-  int currPos[3];
+  float currPos[3];
 
+  // converts angle [deg] to pulse width for the servo driver
   [[nodiscard]] int angle(uint8_t pwmAngle);
 
+  // converts angle [deg] to pulse width for the servo driver
   [[nodiscard]] int angleFloat(float pwmAngle);
 
+  // map() only works for integers
   float mapFloat(float x, float in_min, float in_max, float out_min, float out_max);
 };
